Move main.go to top level
