+++ v2/tools/asoctl/cmd/crd_clean.go
@@ -3,7 +3,7 @@
  * Licensed under the MIT license.
  */
 
-package main
+package cmd
 
 import (
 	"context"
