+++ v2/tools/asoctl/cmd/import_azure_resource.go
@@ -3,7 +3,7 @@
  * Licensed under the MIT license.
  */
 
-package main
+package cmd
 
 import (
 	"context"
