+++ v2/tools/asoctl/main.go
@@ -11,6 +11,8 @@ import (
 
 	flag "github.com/spf13/pflag"
 	"k8s.io/klog"
+
+	"github.com/Azure/azure-service-operator/v2/tools/asoctl/cmd"
 )
 
 func main() {
@@ -18,5 +20,5 @@ func main() {
 	klog.InitFlags(flagSet)
 	flagSet.Parse(os.Args[1:]) //nolint:errcheck // error will never be returned due to ExitOnError
 	flag.CommandLine.AddGoFlagSet(flagSet)
-	Execute()
+	cmd.Execute()
 }
